# Optimize table copy and save step