Another attempt to fix https://github.com/iterative/datachain/issues/220

Initial run:
```
$ time python wal-on.py
Listing file:///: 2 objects [00:00, 750.46 objects/s]
Processed: 2 rows [00:00, 4040.76 rows/s]
Download: 5.78GB [07:36, 13.6MB/s]rows]
Processed: 2 rows [07:36, 228.39s/ rows]
Generated: 1009874 rows [07:36, 2214.38 rows/s]
Processed: 1009874 rows [05:19, 3162.63 rows/s]
python wal-on.py  2707.84s user 439.81s system 81% cpu 1:04:03.64 total
$
```

WAL off:
```
$ time python wal-off.py
Listing file:///: 2 objects [00:00, 687.08 objects/s]
Processed: 2 rows [00:00, 4148.67 rows/s]
Download: 5.78GB [06:10, 16.7MB/s]rows]
Processed: 2 rows [06:10, 185.44s/ rows]
Generated: 1009874 rows [06:10, 2727.49 rows/s]
Processed: 1009874 rows [05:15, 3196.07 rows/s]
python wal-off.py  426.31s user 217.22s system 57% cpu 18:47.79 total
$
```

This PR:
```
$ time python copy_table.py
Listing file:///: 2 objects [00:00, 647.87 objects/s]
Preparing: 2 rows [00:00, 3109.19 rows/s]
Processed: 2 rows [00:00, 4007.94 rows/s]
Download: 5.78GB [07:30, 13.8MB/s]rows]
Processed: 2 rows [07:30, 225.16s/ rows]
Generated: 1009874 rows [07:29, 2245.79 rows/s]
Preparing: 1009874 rows [03:21, 5009.25 rows/s]
Processed: 1009874 rows [04:35, 3671.44 rows/s]
Saving: 1009874 rows [03:09, 5318.70 rows/s]
python copy_table.py  495.25s user 323.62s system 61% cpu 22:08.45 total
$
```

With WAL: `1:04:03`
WAL off: `18:48`
Copy table: `22:08`

More details [below](https://github.com/iterative/datachain/pull/278#issuecomment-2286939542).