@dmpetrov 2024-08-12
@dreadatour please add more details behind the idea. So far, it's not clear why copy is needed and how it can improve performance. Also, please noce, some table are large - WDS 11M table size in ClickHouse is 95Gb.