@dreadatour 2024-08-13
### Details

Test dataset with two files, 1 million rows total:

```
$ ls -lh dvcx-datacomp-small/metadata-big
total 12118496
-rw-r--r--@ 1 vlad  staff   2.9G Jul 24 18:26 0020f0cbd157d470aa56bea08e304b90.npz
-rw-r--r--@ 1 vlad  staff   2.9G Jul 26 10:29 002372587416601d1a339f7589e00276.npz
$
```

Query (`laion_emb_big.py`):

```
from datachain.lib.dc import DataChain
from datachain.lib.webdataset_laion import process_laion_meta

(
    DataChain.from_storage("dvcx-datacomp-small/metadata-big/")
    .gen(emd=process_laion_meta)
    .map(stem=lambda file: file.get_file_stem(), params=["emd.file"], output=str)
    .save("laion_emb")
)
```

Measure query run time:

```
$ time python laion_emb_big.py
Listing file:///: 2 objects [00:00, 750.46 objects/s]
Processed: 2 rows [00:00, 4040.76 rows/s]
Download: 5.78GB [07:36, 13.6MB/s]rows]
Processed: 2 rows [07:36, 228.39s/ rows]
Generated: 1009874 rows [07:36, 2214.38 rows/s]
Processed: 1009874 rows [05:19, 3162.63 rows/s]
python laion_emb_big.py  2707.84s user 439.81s system 81% cpu 1:04:03.64 total
$
```

Total execution time is `1 hour 4 minutes`. All processing/generating steps took 5-7 minutes, so something is going on here. After profiling we can see bottlenecks: [create_pre_udf_table](https://github.com/iterative/datachain/blob/1de7bd322b1e5117553c404df3bf65f4ed23f91b/src/datachain/query/dataset.py#L644) and [save](https://github.com/iterative/datachain/blob/1de7bd322b1e5117553c404df3bf65f4ed23f91b/src/datachain/query/dataset.py#L1661).

Let's measure time:

```
$ time python laion_emb.py
Listing file:///: 2 objects [00:00, 595.61 objects/s]
Created pre-UDF table udf_OAQpEx in 0.00s
Processed: 2 rows [00:00, 4454.92 rows/s]
Download: 5.78GB [07:26, 13.9MB/s]rows]
Processed: 2 rows [07:26, 223.33s/ rows]
Generated: 1009874 rows [07:26, 2264.26 rows/s]
Created pre-UDF table udf_eWKkak in 1389.55s
Processed: 1009874 rows [04:41, 3591.48 rows/s]
python laion_emb.py  2672.79s user 439.16s system 82% cpu 1:02:34.29 total
$
```

`Created pre-UDF table udf_eWKkak` took `1390 seconds`, it is more than `23 minutes`. Let's take a look at executed query:

```sql
INSERT INTO "udf_eWKkak" (
	sys__rand,
	emd__file__source,
	emd__file__path,
	emd__file__size,
	emd__file__version,
	emd__file__etag,
	emd__file__is_latest,
	emd__file__last_modified,
	emd__file__location,
	emd__file__vtype,
	emd__index,
	emd__b32_img,
	emd__b32_txt,
	emd__l14_img,
	emd__l14_txt,
	emd__dedup
)
SELECT 
	anon_1.sys__rand AS sys__rand,
	anon_1.emd__file__source AS emd__file__source,
	anon_1.emd__file__path AS emd__file__path,
	anon_1.emd__file__size AS emd__file__size,
	anon_1.emd__file__version AS emd__file__version,
	anon_1.emd__file__etag AS emd__file__etag,
	anon_1.emd__file__is_latest AS emd__file__is_latest,
	anon_1.emd__file__last_modified AS emd__file__last_modified,
	anon_1.emd__file__location AS emd__file__location,
	anon_1.emd__file__vtype AS emd__file__vtype,
	anon_1.emd__index AS emd__index,
	anon_1.emd__b32_img AS emd__b32_img,
	anon_1.emd__b32_txt AS emd__b32_txt,
	anon_1.emd__l14_img AS emd__l14_img,
	anon_1.emd__l14_txt AS emd__l14_txt,
	anon_1.emd__dedup AS emd__dedup
FROM (
	SELECT
		"udf_N3EBeZ".sys__id AS sys__id,
		"udf_N3EBeZ".sys__rand AS sys__rand,
		"udf_N3EBeZ".emd__file__source AS emd__file__source,
		"udf_N3EBeZ".emd__file__path AS emd__file__path,
		"udf_N3EBeZ".emd__file__size AS emd__file__size,
		"udf_N3EBeZ".emd__file__version AS emd__file__version,
		"udf_N3EBeZ".emd__file__etag AS emd__file__etag,
		"udf_N3EBeZ".emd__file__is_latest AS emd__file__is_latest,
		"udf_N3EBeZ".emd__file__last_modified AS emd__file__last_modified,
		"udf_N3EBeZ".emd__file__location AS emd__file__location,
		"udf_N3EBeZ".emd__file__vtype AS emd__file__vtype,
		"udf_N3EBeZ".emd__index AS emd__index,
		"udf_N3EBeZ".emd__b32_img AS emd__b32_img,
		"udf_N3EBeZ".emd__b32_txt AS emd__b32_txt,
		"udf_N3EBeZ".emd__l14_img AS emd__l14_img,
		"udf_N3EBeZ".emd__l14_txt AS emd__l14_txt,
		"udf_N3EBeZ".emd__dedup AS emd__dedup
	FROM "udf_N3EBeZ"
) AS anon_1
```

Trying to run the same query with same 1M rows table in SQLite CLI — same time (>20 minutes). Even after simplifying this query (removing nested `SELECT`) time is still the same.

Sidenote: after disabling WAL this query took `2-3 minutes` to run.

### Hypotheses

During query run SQLite DB file size and WAL file size growing up quite fast (50-100 Gb). This is because table is huge and each table row is huge. There are a lot of filesystem operations going on:

- this is `select` and `insert` query at the same time
- huge DB file and huge WAL file; SQLite trying to read data from DB **and** from WAL (this is how write-ahead log works)
- at the same time SQLite adding new data to WAL
- and read again

This cause performance bottleneck.

### Solution

Let's make `INSERT FROM SELECT` by chunks. Select all IDs first (IDs only, memory consumption is quite low) and then iterate by list of IDs in batches and run `INSERT FROM SELECT ... WHERE id IN (...)`.

As we see this helps. Performance is almost the same as without WAL.